- Removing custom createSpan code in favor of core-tracing's createSpan functions - Changing messageBatch.tryAdd()'s options with one that matches OperationOptions
