+++ sdk/servicebus/service-bus/review/service-bus.api.md
@@ -11,13 +11,13 @@ import { HttpResponse } from '@azure/core-http';
 import Long from 'long';
 import { MessagingError } from '@azure/core-amqp';
 import { OperationOptions } from '@azure/core-http';
+import { OperationTracingOptions } from '@azure/core-tracing';
 import { PagedAsyncIterableIterator } from '@azure/core-paging';
 import { PageSettings } from '@azure/core-paging';
 import { PipelineOptions } from '@azure/core-http';
 import { RetryMode } from '@azure/core-amqp';
 import { RetryOptions } from '@azure/core-amqp';
 import { ServiceClient } from '@azure/core-http';
-import { Span } from '@opentelemetry/api';
 import { SpanContext } from '@opentelemetry/api';
 import { TokenCredential } from '@azure/core-auth';
 import { TokenType } from '@azure/core-amqp';
@@ -567,7 +567,7 @@ export interface TopicRuntimeProperties {
 
 // @public
 export interface TryAddOptions {
-    parentSpan?: Span | SpanContext | null;
+    tracingOptions?: OperationTracingOptions;
 }
 
 export { WebSocketImpl }
